Move dockerimage into separate package
