+++ pkg/kube/util/operatorimage/docker_image.go
@@ -1,4 +1,4 @@
-package converter
+package operatorimage
 
 import (
 	corev1 "k8s.io/api/core/v1"
